// apps/desktop/src/ipc/generated/StoryBuildArgs.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { StoryBuildDestination } from "./StoryBuildDestination";
import type { StoryBuildInclude } from "./StoryBuildInclude";
import type { StoryDescriptionMode } from "./StoryDescriptionMode";

export type StoryBuildArgs = { episodeId: string | null, include: Array<StoryBuildInclude>, descriptionMode: StoryDescriptionMode, destination: StoryBuildDestination, 
/**
 * Required for NewDraft.
 */
screenplayId: string | null, 
/**
 * Title for a new screenplay document.
 */
title: string | null, draftName: string | null, };